requirements.txt
# Deep Learning
tensorflow==2.21.0

# Data & ML
numpy>=1.26.0
pandas>=2.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0

# Jupyter
notebook>=7.0.0
ipykernel>=6.25.0

# Backend API
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.4.0

# Utilities
tqdm>=4.66.0
matplotlib>=3.7.0